+++ vscode/src/ui/NoteSearcherUi.ts
@@ -10,13 +10,11 @@ export interface NoteSearcherUi {
   currentlyOpenDir: () => string | null;
   promptForSearch: (prefill: string) => Promise<string | undefined>;
   showSearchResults: (files: string[]) => Promise<void>;
-  showRelatedFiles: (files: string[]) => void;
   showNotification: (message: string) => Promise<void>;
   showDeadLinks: (links: Link[]) => Promise<void>;
   showBacklinks: (links: string[]) => Promise<void>;
   notifyIndexingStarted: (indexingTask: Promise<void>) => void;
   showError: (e: Error) => Promise<void>;
-  addCurrentNoteModifiedListener: (listener: FileChangeListener) => void;
   addNoteSavedListener: (listener: FileChangeListener) => void;
   addMovedViewToDifferentNoteListener: (listener: FileChangeListener) => void;
 }